@@ -1,10 +1,8 @@
 package ua.training.admission.entity;
 
+import jakarta.persistence.*;
 import lombok.*;
-import org.hibernate.annotations.OnDelete;
-import org.hibernate.annotations.OnDeleteAction;
 
-import javax.persistence.*;
 import java.util.Set;
 
 @Data
